+++ autokeras/encoders.py
@@ -1,66 +1,10 @@
 import numpy as np
 import tensorflow as tf
 
-from autokeras.engine import serializable
+from autokeras.engine import encoder as encoder_module
 
 
-class Encoder(serializable.Serializable):
-    """Base class for encoders of the prediction targets.
-
-    # Arguments
-        num_classes: Int. The number of classes. Defaults to None.
-    """
-
-    def __init__(self, num_classes=None):
-        self.num_classes = num_classes
-        self._labels = None
-        self._int_to_label = {}
-
-    def fit_with_labels(self, data):
-        """Fit the encoder with all the labels.
-
-        # Arguments
-            data: numpy.ndarray. The original labels.
-        """
-        raise NotImplementedError
-
-    def encode(self, data):
-        """Encode the original labels.
-
-        # Arguments
-            data: numpy.ndarray. The original labels.
-
-        # Returns
-            numpy.ndarray. The encoded labels.
-        """
-        raise NotImplementedError
-
-    def decode(self, data):
-        """Decode the encoded labels to original labels.
-
-        # Arguments
-            data: numpy.ndarray. The encoded labels.
-
-        # Returns
-            numpy.ndarray. The original labels.
-        """
-        raise NotImplementedError
-
-    def get_config(self):
-        return {
-            'num_classes': self.num_classes,
-            'labels': self._labels,
-            'int_to_label': self._int_to_label,
-        }
-
-    @classmethod
-    def from_config(cls, config):
-        obj = super().from_config(config)
-        obj._labels = config['labels']
-        obj._int_to_label = config['int_to_label']
-
-
-class OneHotEncoder(Encoder):
+class OneHotEncoder(encoder_module.Encoder):
     """OneHotEncoder to encode and decode the labels.
 
     This class provides ways to transform data's classification label into vector.
@@ -144,7 +88,7 @@ class OneHotEncoder(Encoder):
                                  np.argmax(np.array(data), axis=1)))).reshape(-1, 1)
 
 
-class LabelEncoder(Encoder):
+class LabelEncoder(encoder_module.Encoder):
     """An encoder to encode the labels to integers.
 
     # Arguments
